requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
cffi==1.17.1
colorama==0.4.6
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.4.0
pytest-django==4.11.1
sqlparse==0.5.3
tzdata==2025.2
Faker
django-extensions
drf-nested-routers
Pillow